pyproject.toml
[project]
name = "your-need"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.5.5",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.2",
    "gradio>=6.10.0",
    "ipykernel>=7.2.0",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.22",
    "langchain-exa>=1.0.0",
    "langchain-sarvam-integration>=0.1.9",
    "langchain-tavily>=0.2.17",
    "pymupdf>=1.27.2.2",
    "python-dotenv>=1.2.2",
    "sarvamai>=0.1.26",
    "sentence-transformers>=5.3.0",
    "streamlit>=1.55.0",
    "unstructured>=0.22.6",
    "uvicorn>=0.42.0",
]
